// extracted product assembly logic from controller
